GraphQL1/requirements.txt
anyio==3.7.1
ariadne==0.20.1
blinker==1.6.2
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
graphql-core==3.2.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.31.0
typing_extensions==4.7.1
Werkzeug==2.3.6
python-dotenv==0.19.2
Flask-Migrate==4.0.4
